fix serverless functionality
