--- src/logic/params.ts
@@ -1,1 +1,0 @@
-export const ipfsGateway = "https://gateway.ipfs.dappnode.io";
